+++ infinigen/assets/objects/creatures/fish.py
@@ -374,7 +374,7 @@ class FishSchoolFactory(BoidSwarmFactory):
             particle_size=U(0.3, 1),
             size_random=U(0.1, 0.7),
             use_rotation_instance=True,
-            lifetime=bpy.context.scene.frame_end - bpy.context.scene.frame_start,
+            lifetime=bpy.context.scene.frame_end - bpy.context.scene.frame_start + 1,
             warmup_frames=1,
             emit_duration=0,  # all particles appear immediately
             emit_from="VOLUME",